firewall-cmd: query mode returns reversed values
The query mode exits the program with exit status 1
when the query is true and with exit status 0 when
the query is false. This should be reversed
For example
firewall-cmd --zone=work --query --service=ssh; echo $?
returns 1 while it should return 0 because ssh service is
enabled (by default) in 'work' zone.