removeInterface() don't require zone
Because interface can always belong only to
one zone it's too strict to require the
zone to be specified. The client also usually
don't care which zone the interface has been
part of, it just wants to tell firewall that
the interface is no longer running.

When removeInterface() is called with empty zone
the interface is now removed from zone it belongs to.