Extension of FirewallClient to support config interface
New classes:
  FirewallClientConfig
  FirewallClientConfigZone
  FirewallClientConfigService
  FirewallClientConfigIcmpType
  FirewallClientConfigZoneSettings
  FirewallClientConfigServiceSettings
  FirewallClientConfigIcmpTypeSettings

Example:

fw = FirewallClient()
for zone in fw.config().listZones():
    print(zone)
    z = fw.config().getZone(zone)
    settings = z.getSettings()
    print(" %s %s" % (z.get_property("name"), settings.getShort()))
    settings.addService("mdns")
    z.update(settings)

Use fw.config() to access the config interface. The settings classes are
providing a simple interface to add and remove settings. Use the settings
classes in the update functions.

All classes provides the same methods as the D-BUS interfaces, also callback
handling for the signals.