Add argument 'name' to Removed signal.
ConfigZone, ConfigService and ConfigIcmpType emit Removed signal
just before the object gets removed from connection.

The problem is that at the moment when receiver gets this Removed
signal, the object which sent it is already unregistered and the
receiver has no way how to get any info (e.g. name) about the object
which sent the signal. The only info that the receiver has is
object path of non-existing object.