firewall-config: catch INVALID_X error when removing zone/service/icmp_type
that has been on the bottom of the list.

How to reproduce:
In persistent mode add zone with name beginning with 'z' letter
and then remove it again.
You'll get INVALID_ZONE error. It's because onRemoveZone() first
tells firewalld to remove the zone, then calls load_zones()
which clears the zoneStore. The clearing evokes onChangeZone()
where we try to get info about the removed zone -> error.