firewall-config: loadDefaults() can throw exception
If the configuration file of zone (or service or icmp type) is stored
in /etc/ it's marked as default even if there's no default configuration
in /usr/lib/ (which is the case of user created zones).
loadDefaults() call for such zone then throws NO_DEFAULT exception.
Therefore we wrap this call in try:except: block.

Note:
*Maybe* it'd be better to mark such a zone as default so firewall-config
would not allow to load the defaults at all.