Firewall.start(): don't call set_default_zone()
This was added with commit 4ec98e3, because
set_default_zone() calls FirewallZone.change_default_zone()
which is needed for adding of "+" (all) interface to default zone.

Problem (even it's quite harmless) is that set_default_zone()
writes the default zone to the configuration file which is
unnecessary, because we've just read it from there.

Therefore we now call directly
FirewallZone.change_default_zone()