Allow to bind interfaces and sources to zones permanently
Zone files can now contain interfaces and source address to bind a zone to.

Example:
<zone>
  <interface name="eth29"/>
  <source family="ipv4" address="192.168.0.0/24"/>
  <source family="ipv6" address="1:2:3:4::"/>
</zone>

The family argument needs to be provided with the source tag. An invertion of
the source is not supported.

The zone of a bound interfaces can be changed in runtime with the command
line tool already, but will be resetted to the persistent setting in the
config file by reload and service restart. The support for persistent changes
will be added later on, altogether with the support for source addresses.
There is also no D-Bus interface for the source handling yet. Also no support
for interfaces and sources in the applet and config tool.