client.handle_exceptions(): do not loop forever
The loop has been there probably to be able to retry
("Retry" button in firewall-config's error dialog)
last function call in case of exception.
In fact in most cases the next try results
in the same exception so it's useless to offer such
option at all.

Moreover up to now the handler has had only
2 possibilities: quit the program or
repeat (forever) the last call.

With this patch we can in the handler for example
ignore the exception and continue the program.