Support splitted zone files
Splitted zones are customized and need to be taken care manually.

Zonei files in /etc/firewalld/zones can be splittet in several files by
using a directory. These zones can not be handled with the config interface
and will be not available there. It is not possible to modify the files using
firewalld.

Example:

/etc/firewalld/zones/external/
                              1.xml
                              2.xml
                              3.xml

If the name matches a zone in /var/lib/firewalld, then this zone will not
be available. There is no way to load defaults for this zone also.