No need to specify whether source address family is IPv4 or IPv6.
FirewallD is able to check this itself, there are functions
checkIPnMask() and checkIP6nMask() for this.

The user as well does not need to see something like:
ipv6:dead:beef::1 or ipv4:1.2.3.0
dead:beef::1 and 1.2.3.0 is sufficient.

FirewallD is still keeping this information internally but does not
pass this in/out.