firewall-cmd: direct options - check ipv4|ipv6|eb (RHBZ#970505)
It's possible to tell (add_argument(choices=[...]))
argparse that argument has to be
selected from set of values, but we can't use it here
because direct options can take more than one argument
(and argparse forces them all to be selected from that set).
So we have to check the value ourselves.