zone's immutable attribute is redundant
We've been marking zone (trusted, block, drop) as immutable when
we didn't want user to change any settings because these zones
have no other chains, only 'jump to <target>' rule.
But it's useless to have immutable attribute because we
know that zone shouldn't be changed when it has other than
default target, i.e. {ACCEPT/REJECT/DROP}.

This commit removes this immutable attribute from zone and
the decision whether zone can be changed is based on the target.

It also allows to overload "immutable" zone during load, because
we need this to permanently add interfaces to "immutable" zones.