firewall-config: twist ICMP Type IP address family logic.
Previously having both IPv4 and IPv6 checkboxes unchecked
and both checked meant the same, i.e. that this ICMP Type
is available for both, IPv4 and IPV6.
That makes the logic obscure.

In ICMP Type (XML) conf. file the destination tag is optional.
In case it's missing it means both IPv4 and IPv6.
So in firewall-config we should in this case actually
show both IPv4 and IPv6 (not none as previously).

Problem is that there's no way how to specify "no IPv4"
and "no IPv6" when empty means all :-)
So this leads into a little weird behaviour - when user
uncheckes both checkboxes (i.e. destination list becomes empty)
both immediatelly become checked (because empty means all, remember ?).
On the other hand trying to specify "no IPv4" and  "no IPv6"
does not make any sense.

I think the best (easy to understand and error prone) solution
would be to make the destination tag mandatory, not optional.