doc: Fixed NONE prefix in man pages
AC_CONFIG_FILES was used tp create doc/xml/translate-*.xsl, but the initial
value for prefix is NONE. Creating the xsl translation files in configure
results in the use of NONE/prefix. Creating the xsl translation files in the
build step, this is not a problem anymore. If configure is used, it removes
the xsl translation files to make sure that the new set variables are used.