Use policy DROP while reloading, do not reset policy in restart twice
Before flushing all rules, the policy of all builtin chains is set to DROP.
After reloading and recreation of all rules and configuration options, the
policy is reset to ACCEPT if panic is not enabled. Otherwise panic is enabled
again.

If panic is not using the builtin chain policy anymore this has to be adapted.