Enforce trust, block and drop zones in the filter table only (RHBZ#1055190)
Add an additional rule with the zone target (accept, reject
or drop) to the base _zone only in the filter table.

Otherwise it is not be possible to have a zone with drop
target, that is allowing traffic that is locally initiated
or that adds additional rules.

Fixed descriptions of block and drop zone to reflect this.