firewall/core/io/*: few improvements (RHBZ#1065738)
Patch from ayanamist

1. When using firewall-cmd --permanent to write rules to direct.xml, the
rule args are not escaped, so when args contain ampersand, it will
become an invalid xml. Use sax.saxutils.escape.

2. When firewalld read and parse direct.xml, it uses
sax.handler.ContentHandler. When the value in a element is very long
characters() will be called several times with
chunks of the data, which overwrites self._element. Use append in
characters(), and add startElement() to reset self._element.