ipXtables: use -w or -w2 if supported
iptables (since 1.4.20) has a locking mechanism [1],
which prevents multiple instances of the program from running concurrently.

It can happen that some other daemon (libvirtd [2], docker [3])
is calling iptables at the same time as firewalld.
In that case the second one which calls iptables fails with:
"Another app is currently holding the xtables lock.
Perhaps you want to use the -w option?"

The easiest work-around is the use the suggested "-w" option,
which makes the second iptables instance wait till the lock is released.
Even better is to use "-w2" [2] which makes it wait for max 2 seconds.

[1] https://git.netfilter.org/iptables/commit/?id=93587a04d0f2511e108bbc4d87a8b9d28a5c5dd8
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1098281
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1151067
[4] https://git.netfilter.org/iptables/commit/?id=aaa4ace72ba1d195bbf436134a336816c33f7bd0