ebtables: use --concurrent to play nicely with other ebtables users (e.g. openstack-nova)
RHBZ#1184782

Similar to iptables' --wait/-w option added with b3b451d6f

Don't check for --concurrent option availability (as we do with --wait),
it has been available since ebtables-2.0.10 (20110710).

With --wait/-w we use -w2 (if available) to limit for how long we wait
if the lock is taken. This was added due to [1].
With --concurrent there shouldn't be such a problem
as it uses file lock in /var/lib/ebtables/.

[1] http://marc.info/?l=netfilter-devel&m=142107406730589&w=2