firewall-applet, firewall-config: Only honour active connections
It is possible to have several connections using the same interface, where
only one connection is active at a time.

firewall-applet and -config are sharing some code to be able to change the
zone of a connection using NetworkManager. The code was taking a look at all
connections and not only the active connections. Because of this it could
happen in rare cases that firewall-applet and firewall-config were trying to
set the new zone for the wrong connection, which then resulted in the
"org.freedesktop.NetworkManager.Settings.PermissionDenied: uid X has no
permission to perform this operation" error if the user was not allowed to
modify the connection.

RHBZ#1190520, RHBZ#1227413