firewall-applet: Fix Ok button sensitivity in dialogs
The Ok buttons in dialogs only should be sensitive, if there have ben changes
in the dialog. The use of setEnabled does not have a good result, but
setDisabled is working as expected.

Also call combo_changed in the end of set_zone.

Dropped unused combo arg in combo_changed.

Call set_zone late in create_ui.