firewall-config: Adapt glade file to newer glade syntax
Do not use H and V versions of Box and Separator, but with additional orientation property.
Use Grid instead of Table.
Drop Alignments where possible and use border_width on the objects.
Some other adaptions.

Due to limitations in glade-3.18.3 not all deprecation warnings are fixable, this version is not providing support for gtk3.13 and later.

gtk3_chooserbutton has been adapted to support Box also.