firewall/core/io/zone: New support for MAC sources
MAC sources can be used in rich rules and also as source bindings for zones.

The use of MAC source bindings are limited: Port forwarding and masquerading
rules in the zones do not have an effect for the MAC source bindings. MAC
source bindings can be used for host black and white listing for example.