ipXtables, ebtables: New support for set_rules methods usind -restore commands
By using the restore commands it is possible to a lot of changes in the firewall in a short amount of time.

The set_rules commands are used in flush and set_policy methods if the individual argument is not set to False.

The new tempFile fuction (firewall/core/functions.py) is used to create the file with the adapted rules for use with the restore comamnds.