firewall/core/fw.py: Use individual calls setting, enable use of restore commands
The use of the restore commands via the Firewalld.rules method is speeding up most actions of firewalld. Only actions that only do single rule changes are not affected by this change.

Now it is possible to have for example a service with thousands of single port entries that is enabled in a second on a reasonable fast machine by firewalld.

Also the start and restart of firewalld are a lot faster with this change.

There is one downside though: The -restore commands are not using xtables_lock or lock_file yet. Therefore the IndividualCalls setting might better default to True for now.