firewall-config: Properly use builtin and default properties of zones, services, ..
In the permanent view, the buttons to remove a zone, service, icmptype and ipset have been sensitive even if there was a builtin for the zone, etc. . It is not possible to remove the zone, etc. it is only possible to load the defaults. LoadDefaults was always sensitive for zones, etc. that are loaded from /etc, even if there was no builtin in /usr/lib.

Additionally it was also allowed in the edit dialog to rename a builtin, which resulted in an error.