FirewallIPSet: Support restore in apply_ipsets, use it in Firewall
The apply_ipsets method has the new argument individual, which derfaults to False. In the individual mode the method behaves as before and uses individual calls for creation of the set and addition of the entries. Otherwise the new restore method in the ipset class is used.

The ipset creation in Firewall has been adapted and has moved right before the creation of zones in netfilter.