Add options --save-default-zone, --get-saved-zone, --restore-saved-zone
In order to save the current default zone for later restore it this
patch adds the options --save-default-zone, --get-saved-zone, --restore-saved-zone.

An use case as example:
A program ships a zone and before setting it as default during
the instalation it could get the current default zone and save it
for when removing the software can restore the previous default zone.