Add options --save-default-zone, --get-saved-zone, --restore-saved-zone
In order to save the current default zone for later restore this patch
adds the options --save-default-zone, --get-saved-zone, --restore-saved-zone.

An use case as example:
A program ships a zone insinde rpm pkg and before setting it as new default zone
during the instalation (%post) it could get the current default zone and save it.
Later when removing the program (%postun) can restore the previous default zone.