src: core: firewalld_conf: Check if IndividualCalls value is supported
Checking if the value for the IndividualCalls option is not empty is not
enough. We also need to check if the supplied value is a supported one
otherwise the value will silently be ignored if it is not 'true' or 'yes'
without feedback to the user.
Finally, refactor IndividualCalls code for consistency with the rest of
the options.