firewall.server.firewalld.FirewallD: New property IPSetTypes, rework of property handling
The IPSetTypes property is a string array, therefore the Get and GetAll methods needed to be reworked to return proper dbus types. An empty array can not be automatically converted to a dbus array, because the type of the array elements is unknown.

The _get_property method is now returning dbus types. The method is used in Get and GetAll. GetAll returnes a dbus.Dictionary now.