Add D-Bus properties to introspection data
firewall.dbus_utils:

- New method dbus_signature
to get the dbus signature type from dbus type. This is used on the return values of the Get(All) methods.

- New method dbus_introspection_prepare_properties
to prepare property information like name, signature and access type name in a dict in the fw_dbus_properties attribute of the dbus.service.Object.

- New method dbus_introspection_add_properties
to add the prepared properties to the Introspect method of the dbus.service.Object.

The Introspect method has been added for all D-Bus interface classes. The properties are added to the XML data gathered from (slip.)dbus.service.Object.