Moved global *tables_available_tables into fw, new get_available_tables method
The global *tables_available_tables in ipXtables and ebtables resulted in a
repeated creation of these tables. The tables have been removed form the
modules. The available tables are now all part of the Firewall class and
generated at initialization. There is the new accessor method
get_available_tables using the ipv argument.