configure.ac: Autodetect the {ip,ip6,eb}tables{,-restore} tools
Assuming a default path for these tools is error-prone since distributions
install these tools into different places and not all packagers make
use of the --with-* switches to set the correct path. As a result of which,
rework the code to use the standard autoconf macros to detect the path for
these tools during the configure phase. Fixes #95