firewall.core.fw_nm: Use NM.client, new nm_get_connection_of_interface
Uses gi.repository.NM in version 1.0 instead of gi.repository.NetworkManager
New function nm_get_connection_of_interface

VPN connections are ignored because they are not handled properly. VPN
connections are using the device of the used connection and there is no zone
assignment. This needs to be implemented in NetworkManager and can use source
addresses for this. Additionally not all devices are assigned to the VPN
connection in NetworkManager like for example tun devices (this should most
likely be fixed in NetworkManager 1.2).

Adapted firewall-config and firewall-applet