firewall-cmd: Try to set the zone of an NetworkManager controlled interface
For --add-interface, --change-interface and --remove-interface NetworkManager
is at first connected to chnage the zone for the connection. If this fails, a
zone binding is created in firewalld.

It is also not possible anymore to remove an interafce zone binding for an
interface that is used for a connection while the connection is still active.