New FirewallD1.config.Methods.getXNames for zones, services, ipsets and icmptypes
With these methods it is a lot faster to get the names of the zones, services, ipsets and icmptypes that are currently available in the permanent environment.

This speeds up switching to the permanent view in firewall-config a lot and also the use of names in firewall-cmd.

The D-Bus interface has been extended for this.
New FirewallClient methods have been added also.