firewall-config: New add/remove ipset entries from file, remove all entries
To simplify ipset handling, there is now support to add and remove enies from
a file to/from an ipset. This is the same as the options
--{add,remove}-entries-from-file for firewall[-offline]-cmd.

The add and remove entry buttons have been replaced with menu buttons.

There is an additional "Remove All Entries" item in the remove menu.