firewall.command: New FirewallCommand for command line client simplification
The output and check functions of firewall-cmd and firewall-offline-cmd have
been merged togeter in the FirewallCommand class. Verbose and quiet options are
also handled by FirewallCommand.

New functionality to handle add, remove and query options as sequences have
been added. These are the methods to handle sequence options:

- add_sequence for things like adding a port to a service or zone in permanent
  configuration.
- x_add_sequence to add entries to an ipset in the runtime configuration.
- zone_add_timeout_sequence to add services to a zone in the runtime
  configruation with an optional timeout setting.
- remove_sequence to remove porst from a service or zone in permanent
  configuration.
- x_remove_sequence to remove entries from an ipset in the runtime
  configuration.
- query_sequence to query if entries are added to an ipset in the permanent
  configuration.
- x_query_sequence to query if entries are added to an ipset in the runtime
  configuration.

These different methods are needed, because the runtime and permanent
configuration interfaces are different. The methods also make sure that the
behaviour with using a single option only is the same as before. If more
options are used, then the output is different as there needs to be a way to
report the setting of all options.