firewall.command: New [de]activate_exception_handler, raise error in parse_X
New [de]activate_exception_handler methods to deactivate the exception_handler
behaviour in the X_sequence methods. If the exception_handler is deactivated
it simply calls raise again.

The parse_method calls in __cmd_sequence and __query_sequence need are now
using exception handling. This makes it possible to test options in the client
before they are tested in add/remove/query calls in firewalld.

The parse_X methods are now raising FirewallErrors instead of calling the fail
method.