config/firewalld.service.in: use network-pre.target
Firewall services should be ordered before network-pre.target and should
pull it, according to :
https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

This avoid any specific and fixed 'Before=' dependencies for
NetworkManager and libvirtd as those services properly order themselves
after network-pre.target or network.target.

Also replace WantedBy=basic.target with multi-user.target as firewalld
is not needed to bootup a system. It only needs to be started before the
network and Before=network-pre.target already covers that.