firewall.core.ipXtables: Split up source and dest addresses for transaction
The use of source and destination address lists is not supported in
ip*tables-restore that is used in transactions. Therefore code has been added
that splits source and destinations addresses into separate rules.

With iptables it is also possible to have source and destination address lists
at the same time.