firewall.command: Fix sequence exit code with at least one succeeded item
For sequences with errors but at least one succeeded item, the exit code should
be 0 as there has been a change. The errors will be printed as warnings.

Without any succeeded item, the exit code will depend on the error codes. If
there is exactly one error code, then this is used. If there are more than one
then UNKNOWN_ERROR (254) will be used.

This is an additional fix for 84f5468 and RHBZ#1366654