firewall.core.fw_transaction: Load helper modules in FirewallZoneTransaction
Helper modules for services that are enabled in the default zone have been
loaded before, but not for zones that are enabled later on because of a
interface/source binding.

This happened because for these zones the FirewallZoneTransaction is used
directly and because of this the added modules are simply not used.