Makefile.am: New dist-check used in the archive target
This check tests if there are config files in the git repo that are not listed
in config/Makefile.am. Also for firewall bindings that are in git, but not in
src/Makefile.am.

This check is not used in the dist tag to be able to create test dists.