firewall.core.{ipXtables,ebtables}: Copy rule before extracting items in set_rules
In set_rules, the rules are grouped by table to be able to create the
iptables-save format without changing the table serveral times.

For this the table is extracted from the rule and therefore also removed from
the rule. But this is not done on a copy of the rule, but the internal rule.
This results in remogin the table information from the rule completely, which
is an issue if the rule can not be applied in the transaction and the
generous_mode is used to be able to assign the rules one by one. This is the
case for rules saved in direct.xml.

Fixes issue #152