firewall.core.io.helper: New IO handler for netfilter helpers
A helper can provide an optional family. The default setting is empty, which
means ipv4 and ipv6. With the family it is possible to limit a helper to ipv4
or ipv6. The irc, netbios-ns, pptp and snmp netfilter helpers are all ipv4 only.

The ports used with the helper can be defined with the port tag in the same way
as for services and zones.