config/firewalld.conf: New AutomaticHelpers setting with description
For the secure use of iptables and connection tracking helpers it is
recommended to turn AutomaticHelpers off. But this might have side effects on
other services using the netfilter helpers as the sysctl setting in
/proc/sys/net/netfilter/nf_conntrack_helper will be changed.
With the system setting, the default value set in the kernel or with sysctl
will be used.
Default: system