firewall.core.io.service: The module does not need to start with nf_conntrack_ anymore
The modules (netfilter helpers) do not need to start with nf_conntrack_ anymore.
If this is stil the case, it will be automatically removed while loading.
Underscores in the module string will also be replaced by dashes automatically.

This is needed to be able to have a consistent helper name usage.