firewall.core.fw_nm: create NMClient lazily
This actually fixes a couple of issues:

NetworkManager.service depends on firewalld.service, so that it doesn't
proceed activating services before it can set the firewall zones. That
means that fw_nm's attempt to create a NMClient always fails.

Secondly, the failure to construct a NMClient is silently discarded in
the except branch. Not cool.

Lastly, asynchronous NMClient creation could be rather expensive if
there's a lot of devices or connections. (Thankfully, it doesn't
succeed, right?). Let's not delay the service startup with it.