firewall.core.io.zone: Fix permanent rich rules using icmp-type (RHBZ#1434594)
Rich language rules using the icmp-type element have not been saved properly.
The code to handle the icmp-type element in the zone writer has been missing
and this has only been logged as a warning. An element without name has been
created because of this. This resulted in a corrupt zone file.

The code to handle the icmp-type element has been added and the warning for
an unknown element has been transformed into a FirewallError. A curruption
of the zone file can not happen anymore with an unhandled element.