firewall.core.fw_icmptype: Add ICMP type only if the type is supported
The icmp type objects is added if the type if the icmp type is in the
supported icmp types list(s) for IPv4 and/or IPv6 according to the icmp
type destination setting.

If an icmp type is only supported by IPv4 or IPv6 and enabled for both, then
a copy of the icmp type will be added with only the supported ipv type. This
only affects run-time and the permanent object is not changed.