Support sctp and dccp in ports, source-ports, forward-ports, helpers and rich rules
This patch adds support to use ports with the protocols sctp and dccp if also
a port id is specified. The use of sctp and dccp is now also allowed in
source-ports, forward-ports, helpers and rich language rules.

The test suite has been expanded to also test the new combinations.

This fixes RHBZ#1429808