src: firewall: Add support for SUSE ifcfg scripts
SUSE stores ifcfg files in /etc/sysconfig/network whereas RedHat stores
them in /etc/sysconfig/network-scripts. These paths can't change so we
can use them to do a 'distribution detection' when looking for an ifcfg
file for the interface we are interested in. Moreover, the SUSE ifcfg
files do not use the DEVICE attribute to denote the interface that is
being configured by a particular ifcfg file. As such, we are only
interested in the ifcfg filename to determine if there is a SUSE ifcfg
file for that interface or not.