Revert "src/firewall/core/ebtabes.py: Remove dangling ebtables lock file"
This reverts commit f431547ffe3369daeec46c781fda68eff7747a81.

It was initially needed because ebtables could leave a stale lockfile,
but now ebtables uses flock() and it's normal for the lock file to
linger. Even the workaround, f431547ffe33, we're reverting here was
insufficient in some cases, see rhbz 1495893.

Now that ebtables does proper locking we can revert this workaround.