backend: introduce nftables support
This introduces a new firewall backend, nftables. It is the new default.
The backend is selectable with FirewallBackend in firewalld.conf. In the
case some currently unsupported features are detected to be in use (e.g.
direct rules) it will fallback to the iptables backend.

Some known issues (still WIP):

    - no direct rules
    - no ipset
    - no AUDIT support