backend: introduce nftables support
This introduces a new firewall backend, nftables. It is the new default.
The backend is selectable with FirewallBackend in firewalld.conf -
possible values are; nftables, iptables.

Some known issues using the nftables backend:

    1) --direct nat rules not allowed by kernel
    2) no ipset in rich rules or zone sources
    3) no AUDIT support (rich rules)

Issue #1 is possible to fix, but requires kernel changes.
Issue #2 is simply not implemented yet.
Issue #3 needs new support in nftables (kernel and userspace).

If the above features are used, then you should revert to using iptables
by setting the following in /etc/firewalld.conf:

   FirewallBackend=iptables

Fixes: #107