src: firewall-config: Fix default value for dialog buttons
The default value of the buttons argument did not work as expected. A
nested tuple with a single item is automatically converted to a regular
tuple and the length() of that tuple was actually the number of items of
that tuple. As a result of which, creating a dialog with a single button
resulted to the following error:

Traceback (most recent call last):
  File "./src/firewall-config", line 1623, in connection_failed
    message_type=Gtk.MessageType.ERROR, msg=msg) == 1:
  File "./src/firewall-config", line 1596, in _dialog
    for button,id in buttons:
ValueError: too many values to unpack

The way to solve that, is to convert the nested tuple to a list of
tuples instead.