src: firewall-config: Break infinite loop when firewalld is not running
If firewalld is not running then the firewall-config UI will wait
forever until it manages to establish a dbus connection with the
service. However, this is not optimal since it provides no feedback to
the user as to what it is waiting for. It's best to use a reasonable
timeout value like 15 seconds which should be more than enough for the
connection be established. If that does not happen, then we print a
useful message in order for the user to start investigating the reason
behind the failure.

Fixes #299