Revert "travis: run testsuite in parallel"
This reverts commit f59ab59c86ed3d231b27ab3cc8db32f3353ff6ad.

Ubuntu Trusty still seems to have issue with concurrent ebtables
invocations (even with --concurrent). Even though we're running the
tests is separate namespaces the ebtables lock is still just a normal
file. The locking was fixed to use flock() in newer ebtables versions,
but that fix doesn't appear to be in Trusty.