travis-ci: use pip to install python packages
It was found that due to hardcoded interpreters firewalld was always
using python2. This is true even in travis-ci virtualenvs.

As such, install dependencies from pip so we get the correct versions
for each virtualenv. Some dependencies we still have to get from apt
(gi, dbus). This limits the versions of python we can actually test;
2.7, 3.4.