core: Replace 'icmp' with ipv6-icmp for IPv6 iptables rules
When we select the protocol from the UI, the value is being taken from
the drop-down combobox. The combobox only lists 'icmp' to cover both
icmp and ipv6-icmp protocols but this results to icmp being injected
into the IPv6 chains so it does not work as expected. The solution to
that would be to add 'ipv6-icmp' protocol to cover the IPv6 case.

Link: https://bugzilla.suse.com/show_bug.cgi?id=1099698