core: logger: Remove world-readable bit from logfile
The firewalld logfile was created with the 'read' bit set for all users
which could potentially reveal information about the firewall
configuration to unauthorised users. As such, we need to make use of
the low-level I/O facilities of the 'os' module to drop that bit during
file opening and creation.

Co-Authored-By: Matthias Gerstner <matthias.gerstner@suse.com>