firewall-cmd: On getZoneOfInterface only ask NM for permanent conns
When the --permanent flag is set first ask NetworkManager if the
interface is in it's permanent configuration. Then we can fallback to
firewalld's config. Previously this just went straight to firewalld's
config which was usually empty in NetworkManger was in use.

Fixes: rhbz 1460295
(cherry picked from commit 02b194ee3fbf11f1dcb406565ef8cf53a2fc74b9)