New firewalld configuration setting DropOnReload
Currently, while reloading the policy for built-in chains is set to
DROP for all the tables and is then reset to ACCEPT at the end.
Because the policy of the raw/mangle/security table is set to DROP
and lack of ACCEPT rules, connection interruption is possible during
the reload.

If the user think the default rules in the filter table are secure,
the new DropOnReload could be set as no to avoid service disruption.

Default: yes