core: Allow icmpv6 ND messages in the filter table
Previously, these messages were only accepted in the raw table. However,
when something is accepted in a netfilter table, it skips the rest of
the chains but it still traverses the rest of the tables.

Moreover, when using nftables, it still traverses chains with higher
priority on the same hook.

So these icmpv6 packets were accepted in the 'raw' table but they still
went through the mangle and filter tables. When a table uses DROP/REJECT
as default policy (eg drop zone) then these messages are eventually
dropped in the filter table so ND stops working. As such, we need to
accept these messages in that table.