ICMP_BLOCK rule index incorrect when logging is on
When LogDenied is enabled, logging rules created for ICMP BLOCKS are
incorrectly indexed for both IPv4 and IPv6, leading to a the restore tools
for each failing to load the new rules. This was caused by the rule_idx
variable incorrectly being reset in the outer loop instead of the inner loop.