firewall: core: Only apply previous runtime configuration on startup failure
Commit 2796edc1691f ("fw: if startup fails on reload, reapply non-perm
config that survives reload") made the firewalld code more resilient to
startup failures by re-applying the runtime configuration if firewalld
startup method failed. However, this only must be done in case of
failures and not all the time since this means that permanent
configuration is not loaded properly as demonstrated below:

 firewall-cmd --zone=home --change-interface=eth0
 success
 firewall-cmd --zone=home --list-interfaces
 eth0
 firewall-cmd --permanent --zone=public --change-interface=eth0
 success
 firewall-cmd --list-interfaces --zone=home
 eth0
 firewall-cmd --reload
 firewall-cmd --list-interfaces --zone=public
 firewall-cmd --list-interfaces --zone=home
 eth0

As we can see, the runtime configuration of the 'public' zone is not
loaded properly and eth0 still remains assigned to the 'home' zone.

Fixes: suse bz #1109153
Fixes: 2796edc1691f ("fw: if startup fails on reload, reapply non-perm config that survives reload")